requirements.txt
pysphero>=0.1.5
opencv-python>=4.5.0
numpy>=1.19.0
pyyaml>=5.4.0 